Payroll export: split CSV onto its own resource route
The .tsx route is nested under the app layout, so when its loader
returned a text/csv Response React Router still rendered the layout
shell around it — the download arrived as HTML wrapping the CSV body
instead of a bare CSV stream.

Splits the CSV out to admin-console.payroll-export.csv.ts (no default
export, registered outside the layout block) so the Response streams
as-is. Shared row-building + RFC 4180 quoting move to
~/admin-console/lib/payroll-export.ts; the page route imports the same
helpers for its preview table so the preview and the download stay
guaranteed-identical.

Download button is now a plain link to /admin-console/payroll-export.csv
rather than a ?format=csv variant of the page URL.
