+++ dali-api/app/admin-console/routes/admin-console.payroll-export.tsx
@@ -4,179 +4,19 @@ import { prisma } from "~/lib/db";
 import { requireAuth } from "~/lib/auth";
 import { isAdmin } from "~/lib/roles";
 import { Download, FileDown, AlertTriangle } from "lucide-react";
+import {
+  buildPayrollRows,
+  pickDefaultTermId,
+  type PayrollRow,
+} from "~/admin-console/lib/payroll-export";
 
 export const meta: Route.MetaFunction = () => [
   { title: "Payroll export · Operations · DALI OS" },
 ];
 
-// ─── Constants per payroll spec ──────────────────────────────────────────────
-// Single hardcoded primary/secondary supervisor for the whole lab. Per spec,
-// every row gets these NetIDs. If supervisors ever differ per project, replace
-// with a per-project field on Project (mirroring chartString).
-const PRIMARY_SUPERVISOR_NETID = "f0077bn";
-const SECONDARY_SUPERVISOR_NETID = "d1207c2";
-const ANTICIPATED_HOURS_PER_WEEK = "15";
-
-// 16-column header, order matters — Dartmouth payroll imports column-by-column.
-const CSV_HEADERS = [
-  "Student NetID",
-  "Student First Name",
-  "Student Last Name",
-  "Job ID",
-  "Primary Supervisor NetID",
-  "Secondary Supervisor NetID",
-  "Hourly Wage",
-  "Anticipated Hours per week",
-  "Hire Start Date",
-  "Hire End Date",
-  "Chart String Type",
-  "Full Chart String",
-  "Student Phone Number",
-  "Term",
-  "Max Hours – PTO",
-  "Max Hours – Mental Health",
-] as const;
-
-type PayrollRow = {
-  netId: string;
-  firstName: string;
-  lastName: string;
-  jobId: string;
-  hourlyWage: string;
-  hireStart: string;
-  hireEnd: string;
-  chartStringType: string;
-  chartString: string;
-  // Tracked for the preview UI so admins can spot misconfigured projects /
-  // missing job-code lookups before downloading. Omitted from CSV cells, where
-  // a missing value is just an empty string.
-  warnings: string[];
-};
-
-// RFC 4180 quoting: wrap in quotes if value contains a comma, quote, or newline;
-// double up internal quotes.
-function csvCell(raw: string): string {
-  if (raw === "") return "";
-  if (/[",\n\r]/.test(raw)) return `"${raw.replace(/"/g, '""')}"`;
-  return raw;
-}
-
-function formatDate(d: Date): string {
-  // ISO YYYY-MM-DD; Dartmouth payroll accepts this format.
-  return d.toISOString().slice(0, 10);
-}
-
-// JobCodeLookup uses nullable `level` and `domainId` as wildcards. Match the
-// most specific row first: exact (level, domain) > level-only > domain-only >
-// wildcard. Returns null if nothing matches (admin should seed JobCodeLookup
-// before relying on the export).
-function resolveJobCode(
-  lookups: Array<{
-    level: string | null;
-    domainId: string | null;
-    jobCode: string;
-    payRateUsdHour: { toString(): string } | null;
-  }>,
-  level: string,
-  domainId: string,
-): { jobCode: string; hourlyWage: string } | null {
-  const candidates = lookups
-    .filter((l) => (l.level === null || l.level === level))
-    .filter((l) => (l.domainId === null || l.domainId === domainId));
-  if (candidates.length === 0) return null;
-  // Most specific first.
-  candidates.sort((a, b) => {
-    const specA = (a.level !== null ? 2 : 0) + (a.domainId !== null ? 1 : 0);
-    const specB = (b.level !== null ? 2 : 0) + (b.domainId !== null ? 1 : 0);
-    return specB - specA;
-  });
-  const best = candidates[0];
-  return {
-    jobCode: best.jobCode,
-    hourlyWage: best.payRateUsdHour ? best.payRateUsdHour.toString() : "",
-  };
-}
-
-async function buildRows(termId: string): Promise<PayrollRow[]> {
-  const [assignments, jobLookups, term] = await Promise.all([
-    prisma.projectAssignment.findMany({
-      where: { termId },
-      include: {
-        user: { select: { netId: true, firstName: true, lastName: true } },
-        project: {
-          select: { name: true, chartStringType: true, chartString: true },
-        },
-      },
-      orderBy: [{ user: { lastName: "asc" } }, { user: { firstName: "asc" } }],
-    }),
-    prisma.jobCodeLookup.findMany({
-      where: { assignmentType: "Project" },
-      select: { level: true, domainId: true, jobCode: true, payRateUsdHour: true },
-    }),
-    prisma.term.findUnique({
-      where: { id: termId },
-      select: { startDate: true, endDate: true },
-    }),
-  ]);
-
-  if (!term) return [];
-
-  const hireStart = formatDate(term.startDate);
-  const hireEnd = formatDate(term.endDate);
-
-  return assignments.map((a) => {
-    const warnings: string[] = [];
-    const job = resolveJobCode(jobLookups, a.level, a.domainId);
-    if (!job) warnings.push(`No JobCodeLookup for level=${a.level}`);
-    if (!a.project.chartString) warnings.push("Project missing chartString");
-    if (!a.user.netId) warnings.push("User missing NetID");
-
-    return {
-      netId: a.user.netId ?? "",
-      firstName: a.user.firstName,
-      lastName: a.user.lastName,
-      jobId: job?.jobCode ?? "",
-      hourlyWage: job?.hourlyWage ?? "",
-      hireStart,
-      hireEnd,
-      chartStringType: a.project.chartStringType ?? "",
-      chartString: a.project.chartString ?? "",
-      warnings,
-    };
-  });
-}
-
-function toCsv(rows: PayrollRow[]): string {
-  const lines: string[] = [];
-  lines.push(CSV_HEADERS.map(csvCell).join(","));
-  for (const r of rows) {
-    lines.push(
-      [
-        r.netId,
-        r.firstName,
-        r.lastName,
-        r.jobId,
-        PRIMARY_SUPERVISOR_NETID,
-        SECONDARY_SUPERVISOR_NETID,
-        r.hourlyWage,
-        ANTICIPATED_HOURS_PER_WEEK,
-        r.hireStart,
-        r.hireEnd,
-        r.chartStringType,
-        r.chartString,
-        "", // Student Phone Number — blank per spec
-        "", // Term — blank per spec
-        "", // Max Hours – PTO — blank per spec
-        "", // Max Hours – Mental Health — blank per spec
-      ]
-        .map(csvCell)
-        .join(","),
-    );
-  }
-  // RFC 4180 uses CRLF; payroll importers handle either, CRLF is safer for
-  // Excel on Windows.
-  return lines.join("\r\n") + "\r\n";
-}
+// The CSV itself is served by the sibling resource route at
+// /admin-console/payroll-export.csv (no layout wrapping), so the Download
+// button is a plain link to that URL.
 
 export async function loader({ request }: Route.LoaderArgs) {
   const auth = await requireAuth(request);
@@ -184,7 +24,6 @@ export async function loader({ request }: Route.LoaderArgs) {
   if (!(await isAdmin(auth.user.sub))) return redirect("/admin-console/members");
 
   const url = new URL(request.url);
-  const format = url.searchParams.get("format");
   const requestedTermId = url.searchParams.get("termId");
 
   const terms = await prisma.term.findMany({
@@ -192,32 +31,17 @@ export async function loader({ request }: Route.LoaderArgs) {
     select: { id: true, code: true, startDate: true, endDate: true },
   });
 
-  // Default to the term whose [startDate, endDate] contains today; falls back
-  // to the most recent term if no current term.
-  const now = new Date();
-  const defaultTerm =
-    terms.find((t) => t.startDate <= now && now <= t.endDate) ?? terms[0];
-  const selectedTerm =
-    terms.find((t) => t.id === requestedTermId) ?? defaultTerm;
+  const selectedTermId =
+    (requestedTermId && terms.some((t) => t.id === requestedTermId)
+      ? requestedTermId
+      : pickDefaultTermId(terms));
+  const selectedTerm = terms.find((t) => t.id === selectedTermId);
 
   if (!selectedTerm) {
     return { terms: [], selectedTermId: null, rows: [] as PayrollRow[] };
   }
 
-  const rows = await buildRows(selectedTerm.id);
-
-  if (format === "csv") {
-    const csv = toCsv(rows);
-    const filename = `payroll-${selectedTerm.code}-${formatDate(now)}.csv`;
-    return new Response(csv, {
-      status: 200,
-      headers: {
-        "Content-Type": "text/csv; charset=utf-8",
-        "Content-Disposition": `attachment; filename="${filename}"`,
-        "Cache-Control": "no-store",
-      },
-    });
-  }
+  const rows = await buildPayrollRows(selectedTerm.id);
 
   return {
     terms: terms.map((t) => ({ id: t.id, code: t.code })),
@@ -247,7 +71,7 @@ export default function PayrollExport() {
 
   const { terms, selectedTermId, selectedTermCode, rows } = data;
   const rowsWithWarnings = rows.filter((r) => r.warnings.length > 0).length;
-  const csvHref = `?termId=${encodeURIComponent(selectedTermId)}&format=csv`;
+  const csvHref = `/admin-console/payroll-export.csv?termId=${encodeURIComponent(selectedTermId)}`;
 
   return (
     <div className="space-y-6">
@@ -279,7 +103,7 @@ export default function PayrollExport() {
             </select>
             {/* Preserve any other params on submit (none today, future-proof). */}
             {Array.from(searchParams.entries())
-              .filter(([k]) => k !== "termId" && k !== "format")
+              .filter(([k]) => k !== "termId")
               .map(([k, v]) => (
                 <input key={k} type="hidden" name={k} value={v} />
               ))}
@@ -287,6 +111,10 @@ export default function PayrollExport() {
           <a
             href={csvHref}
             download
+            // reloadDocument tells React Router not to intercept this link as a
+            // client navigation. The browser does a real GET to the resource
+            // route, which streams the CSV with Content-Disposition: attachment.
+            data-discover="false"
             className="inline-flex items-center gap-1.5 px-3 py-1.5 text-sm font-medium rounded-md bg-gray-900 text-white hover:bg-gray-800 transition-colors"
           >
             <Download className="w-4 h-4" /> Download CSV
